tests: Fix/add unit tests based on addition of `config_manipulator.py`
